changing the datatype of instruction latency from uint32_t to uint64_t to prevent loss of bits

+++ src/tool/hpcrun/gpu/instrumentation/kernel-data.h
@@ -31,7 +31,7 @@ typedef struct kernel_data_gtpin_inst {
   float W_ins;
   bool isPredictable;
   bool isComplex;
-  uint32_t aggregated_latency;
+  uint64_t aggregated_latency;
   struct kernel_data_gtpin_inst *next;
 } kernel_data_gtpin_inst_t;
